+++ contact_picker_platform_interface/lib/src/exceptions.dart
@@ -1,0 +1,6 @@
+/// Exception thrown when a user hits the back button without selecting a contact
+class UserCancelledPickingException implements Exception {
+  @override
+  String toString() =>
+      'UserCancelledPickingException: The user hit the back button instead of selecting a contact';
+}