+++ contact_picker_platform_interface/lib/src/method_channel_contact_picker.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:contact_picker_platform_interface/contact_picker_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 import 'email_contact.dart';
@@ -12,25 +13,36 @@ const MethodChannel _channel = const MethodChannel('me.schlaubi.contactpicker');
 class MethodChannelContactPicker extends ContactPickerPlatform {
   @override
   Future<PhoneContact> pickPhoneContact({bool askForPermission = true}) async =>
-      PhoneContact.fromMap((await (_channel.invokeMethod<Map<dynamic, dynamic>>(
-              'pickPhoneContact', {'askForPermission': askForPermission})
-          as FutureOr<Map<dynamic, dynamic>>)));
+      PhoneContact.fromMap((await (_runCancellable(_channel
+          .invokeMethod<Map<dynamic, dynamic>>(
+              'pickPhoneContact', {'askForPermission': askForPermission})))));
 
   @override
   Future<EmailContact> pickEmailContact({bool askForPermission = true}) async =>
-      EmailContact.fromMap((await (_channel.invokeMethod<Map<dynamic, dynamic>>(
-              'pickEmailContact', {'askForPermission': askForPermission})
-          as FutureOr<Map<dynamic, dynamic>>)));
+      EmailContact.fromMap((await (_runCancellable(_channel
+          .invokeMethod<Map<dynamic, dynamic>>(
+              'pickEmailContact', {'askForPermission': askForPermission})))));
 
   @override
   Future<FullContact> pickFullContact({bool askForPermission = true}) async =>
-      FullContact.fromMap((await (_channel.invokeMethod<Map<dynamic, dynamic>>(
-              'pickContact', {'askForPermission': askForPermission})
-          as FutureOr<Map<dynamic, dynamic>>)));
+      FullContact.fromMap((await (_runCancellable(_channel
+          .invokeMethod<Map<dynamic, dynamic>>(
+              'pickContact', {'askForPermission': askForPermission})))));
 
   @override
   Future<bool> hasPermission() async =>
       (await _channel.invokeMethod('hasPermission'));
+
+  Future<T> _runCancellable<T>(FutureOr<T?> cancellable) async {
+    try {
+      return (await cancellable)!;
+    } on PlatformException catch (e) {
+      if (e.code == 'CANCELLED') {
+        throw UserCancelledPickingException();
+      }
+      rethrow;
+    }
+  }
 
   @override
   Future<bool> requestPermission({bool force = false}) async {