+++ contact_picker_web/lib/src/exceptions.dart
@@ -5,10 +5,3 @@ class PickerNotAvailableException implements Exception {
   String toString() =>
       'PickerNotAvailableException: The web contact picker API is not available in this browser. Right now it is only available in mobile versions of some chromium browser. Read this for more info: https://web.dev/contact-picker/ && https://developer.mozilla.org/en-US/docs/Web/API/Contact_Picker_API#Browser_compatibility';
 }
-
-/// Exception thrown when a user hits the back button without selecting a contact
-class UserCancelledPickingException implements Exception {
-  @override
-  String toString() =>
-      'UserCancelledPickingException: The user hit the back button instead of selecting a contact';
-}