+++ contact_picker/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.4.0
+- Fix pickEmailContact never completes if pressing BACK on Android bug (#29)
+
+# 4.3.3
+- Fix pickEmailContact never completes if pressing BACK on Android bug (#29)
+
 # 4.3.1
 - Fix SDK constraints